# front/src/app/api/admin/orders/[id]/route.ts
import { NextRequest, NextResponse } from "next/server";

const API_BASE_URL = process.env.NEXT_PUBLIC_API_URL || "http://localhost:8000";

export async function PATCH(
  req: NextRequest,
  { params }: { params: { id: string } }
) {
  try {
    const { id } = params;
    const body = await req.json();

    const response = await fetch(`${API_BASE_URL}/api/admin/orders/${id}`, {
      method: "PATCH",
      headers: {
        "Content-Type": "application/json",
        Cookie: req.headers.get("cookie") || "",
      },
      body: JSON.stringify(body),
    });

    if (!response.ok) {
      const errorText = await response.text();
      console.error("Error updating order:", errorText);
      return NextResponse.json(
        { error: `HTTP error! status: ${response.status}` },
        { status: response.status }
      );
    }

    const updatedOrder = await response.json();
    return NextResponse.json(updatedOrder);
  } catch (e: unknown) {
    console.error("Error updating order:", e);
    return NextResponse.json(
      {
        error: "Failed to update order",
        details: e instanceof Error ? e.message : String(e),
      },
      { status: 500 }
    );
  }
}